prompt restore

+++ mediawiki/backup/db-restore.sh
@@ -6,11 +6,19 @@ source ${SOURCE_DIR}/env.sh
 CONTAINER="${NAME}-mariadb-${MEDIAWIKI_VF}"
 DATE=`date +%Y-%m-%d--%H-%M-%S`
 BACKUP_DIR="${SCRIPT_PATH}/db"
-LAST_BACKUP="${BACKUP_DIR}/`ls -r1 ${BACKUP_DIR} | grep -E '.sql$' | head -n 1`"
+LAST_FILE="`ls -r1 ${BACKUP_DIR} | grep -E '.sql$' | head -n 1`"
+LAST_BACKUP="${BACKUP_DIR}/${LAST_FILE}"
 
 echo "Restore Database:"
 echo ""
-echo "Last backup: ${LAST_BACKUP}"
+echo "Last backup: ${LAST_FILE}"
+
+read -p "Are you sure to restore this backup ${LAST_FILE}? y/n: " input
+
+if [[ "$input" != "y" ]]; then
+    echo "Aborted"
+    exit 0
+fi
 
 ${COMMAND} exec -i ${CONTAINER} mariadb -u root -e "DROP DATABASE ${NAME};"
 ${COMMAND} exec -i ${CONTAINER} mariadb -u root < ${LAST_BACKUP}
